requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
idna==3.7
pillow==10.4.0
psycopg-binary==3.2.1
psycopg2==2.9.9
pyTelegramBotAPI==4.21.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
setuptools==72.1.0
sqlparse==0.5.1
telebot==0.0.5
urllib3==2.2.2
